neues main file

+++ stereostrand_main.py
@@ -2,6 +2,8 @@ import time
 import numpy as np
 from midi_connector import midi_connector
 
+from pygame import mixer # Load the required library
+
 from asciimatics.effects import Cycle, Stars
 from asciimatics.renderers import FigletText
 from asciimatics.scene import Scene
@@ -25,16 +27,26 @@ def ascii_greeter(screen):
 
 WINDOW = 100
 
-CHANNELS = 9
+# CHANNELS = 6
 
-INTERVAL = int(input('\n\nZeitintervall zwischen den Sounds, z.B. 2 heißt alle 2 sek. einen Ton\n'))
+PATH_TO_MP3 = ''
 
-BEATS_INTERVAL = int(input('\n\nWieviele Minuten zwischen den Beats?\n'))
+INTERVAL = int(input('\n\nZeitintervall zwischen den Sounds, z.B. 30 heißt alle 30 min spielts (nur ganze Zahlen, kein Komma!)\n'))
+
+PLAY_TIME = int(input('\n\nWie lange solls spielen (Minuten)?\n'))
+
+MIN_SAMPLE_INTERVAL = float(input('\n\nWie lang soll der MINIMALE Zeitabstand der samples sein? zB 0.2 für 0.2 sek\n'))
+
+MAX_SAMPLE_INTERVAL = float(input('\n\nWie lang soll der MAXIMALE Zeitabstand der samples sein? zB 1 für 1 sek\n'))
+
+SAMPLE_INTERVAL_RATIO = int(MAX_SAMPLE_INTERVAL/MIN_SAMPLE_INTERVAL)
+
+CHANNELS = int(input('\n\nWie viele Midi Kanäle gibts? (8 heißt 8 Kanäle!)\n'))
 
 # BEATS_INTERVAL in sec
-BEATS_INTERVAL=BEATS_INTERVAL*60
+INTERVAL=INTERVAL*60
 
-LIEBLINGS_ZAHL = int(input('\n\nWas ist deine Lieblingszahl (1-1000)?\n'))
+PLAY_TIME = PLAY_TIME*60/MIN_SAMPLE_INTERVAL
 
 COUNTER = 0
 #
@@ -58,41 +70,52 @@ except:
 
 midi = midi_connector(WINDOW=WINDOW,CHANNELS=CHANNELS)
 
+print('\nIt`s on!')
+print('Press Ctrl+c to abort')
+
+
+
 def play():
 
-    print('\nIt`s on!')
-    print('Press Ctrl+c to abort')
+    counter = 0
 
     # StartZeit
     time_start = time.time()
+
+    # initialize midi
+    midi.init_midi()
+
+    # load the mp3 sound file
+    # initialze the mixer for mp3
+    mixer.init()
+    mixer.music.load(PATH_TO_MP3)
+    # mixer.music.play()
     try:
-        while True:
-            midi.send_signal()
+        while counter<PLAY_TIME:
 
-            time.sleep(INTERVAL * np.random.rand() + INTERVAL)
+            # Das Signal soll random zwischen MIN und MAX gesendet werden
+            my_rand_int = np.random.randint(SAMPLE_INTERVAL_RATIO)
+            if my_rand_int == 0:
+                midi.send_signal()
 
-            RANDINT = np.random.randint(1000)
-
-            #
-            if RANDINT == LIEBLINGS_ZAHL:
-                print('Checkpot!')
-                for i in range(9):
-                    midi.test_send(i)
-                    time.sleep(0.1)
+            time.sleep(MIN_SAMPLE_INTERVAL)
+            counter +=1
 
             # aktuelle Zeit
             jetzt_zeit=time.time()
             diff_time = jetzt_zeit - time_start
 
-            if diff_time > BEATS_INTERVAL:
-                print('\nTHE BEAT IS ON!\n')
-                midi.test_send(9)       # schick signal an Kanal 10 (ist in python die nr 9)
-                time_start = jetzt_zeit
+        # close midi connection
+        midi.quit_midi()
 
     except KeyboardInterrupt:
         midi.quit_midi()
 
-if __name__=='__main__':
-    play()
-    midi.quit_midi()
+
+#TODO
+# hier brauchts noch ein scheduler der alle 30 min play macht
+
+# if __name__=='__main__':
+#     play()
+#     midi.quit_midi()
